requirements.txt
fastapi
uvicorn
google-genai
google-cloud-speech
google-cloud-texttospeech
pydantic
python-dotenv
pytest
httpx
